+++ src/main/java/net/brnbrd/delightful/common/item/knife/compat/undergarden/CloggrumKnifeItem.java
@@ -12,6 +12,6 @@ public class CloggrumKnifeItem extends CompatKnifeItem {
 
 	@Override
 	public String[] getConflicts() {
-		return new String[]{"undergardendelight"};
+		return new String[]{ Mods.UGD };
 	}
 }