+++ src/main/java/net/brnbrd/delightful/common/item/knife/compat/undergarden/ForgottenKnifeItem.java
@@ -34,7 +34,7 @@ public class ForgottenKnifeItem extends CompatKnifeItem {
 
 	@Override
 	public String[] getConflicts() {
-		return new String[]{"undergardendelight"};
+		return new String[]{ Mods.UGD };
 	}
 
 	@Override