+++ src/main/java/net/brnbrd/delightful/common/item/knife/compat/undergarden/FroststeelKnifeItem.java
@@ -21,7 +21,7 @@ public class FroststeelKnifeItem extends CompatKnifeItem {
 
 	@Override
 	public String[] getConflicts() {
-		return new String[]{"undergardendelight"};
+		return new String[]{ Mods.UGD };
 	}
 
 	@Override