+++ src/main/java/net/brnbrd/delightful/common/item/knife/compat/undergarden/UtheriumKnifeItem.java
@@ -23,7 +23,7 @@ public class UtheriumKnifeItem extends CompatKnifeItem {
 
 	@Override
 	public String[] getConflicts() {
-		return new String[]{"undergardendelight"};
+		return new String[]{ Mods.UGD };
 	}
 
 	@Override