+++ src/main/java/net/brnbrd/delightful/compat/JEIPlugin.java
@@ -16,7 +16,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import org.jetbrains.annotations.NotNull;
 import vectorwing.farmersdelight.common.utility.TextUtils;
@@ -32,33 +31,22 @@ public class JEIPlugin implements IModPlugin
     private static final ResourceLocation ID = Util.rl(Delightful.MODID, "jei_plugin");
     private static final Minecraft MC = Minecraft.getInstance();
 
-    private void remove(List<ItemStack> removalList, String modid, String item) {
-        if (!Mods.loaded(modid)) {
-            return;
-        }
-        ResourceLocation location = Util.rl(modid, item);
-        if (ForgeRegistries.ITEMS.containsKey(location)) {
-            Item found = ForgeRegistries.ITEMS.getValue(location);
-            if (found != null) {
-                removalList.add(found.getDefaultInstance());
-            }
-        }
-    }
-
     @Override
     public void registerRecipes(IRecipeRegistration registration) {
-        // Remove all disabled Items from JEI
-        List<ItemStack> removal = new ArrayList<>(DelightfulItems.ITEMS.getEntries().stream()
+        // Hide all disabled items from JEI
+        List<ItemStack> hidden = new ArrayList<>(DelightfulItems.ITEMS.getEntries()
+            .stream()
             .filter(i -> (i.get() instanceof IConfigured c) ? !c.enabled() : !Util.enabled(i)) // Keep items not enabled
             .map(Util::gs) // Get ItemStack
             .toList());
-        this.remove(removal, Mods.AA, "fried_egg");
-        this.remove(removal, Mods.IN, "fried_egg");
-        this.remove(removal, Mods.MD, "bread_slice");
-        this.remove(removal, Mods.MD, "toast");
-        this.remove(removal, Mods.VD, "pb_j");
-        if (removal.size() > 0) {
-            registration.getIngredientManager().removeIngredientsAtRuntime(VanillaTypes.ITEM_STACK, removal);
+        this.hide(hidden, Mods.AA, "fried_egg");
+        this.hide(hidden, Mods.IN, "fried_egg");
+        this.hide(hidden, Mods.MD, "bread_slice");
+        this.hide(hidden, Mods.MD, "toast");
+        this.hide(hidden, Mods.VD, "pb_j");
+        this.hide(hidden, Mods.UGD, "gloomgourd_pie_slice");
+        if (hidden.size() > 0) {
+            registration.getIngredientManager().removeIngredientsAtRuntime(VanillaTypes.ITEM_STACK, hidden);
         }
 
         // Add Knife translations
@@ -104,6 +92,16 @@ public class JEIPlugin implements IModPlugin
         }
         registration.addIngredientInfo(Items.MELON.getDefaultInstance(), VanillaTypes.ITEM_STACK, Component.translatable(Delightful.MODID + ".sliceable.desc"));
         registration.addIngredientInfo(Items.PUMPKIN.getDefaultInstance(), VanillaTypes.ITEM_STACK, Component.translatable(Delightful.MODID + ".sliceable.desc"));
+    }
+
+    private void hide(List<ItemStack> hiddenList, String modid, String item) {
+        if (Mods.loaded(modid)) {
+            ResourceLocation location = Util.rl(modid, item);
+            Item found = Util.item(location);
+            if (found != null) {
+                hiddenList.add(found.getDefaultInstance());
+            }
+        }
     }
 
     @Override