+++ src/main/java/net/brnbrd/delightful/data/tags/DelightfulItemTags.java
@@ -117,10 +117,11 @@ public class DelightfulItemTags {
 	public static final TagKey<Item> TOOLS_MACHETES = forge("tools/machetes");
 	public static final TagKey<Item> SEEDS_SALMONBERRY = forge("seeds/salmonberry");
 	public static final TagKey<Item> SEEDS_CANTALOUPE = forge("seeds/cantaloupe");
+	public static final TagKey<Item> EGGS_BIRD = forge("eggs/bird");
+	public static final TagKey<Item> EGGS_DUCK = forge("eggs/duck");
+	public static final TagKey<Item> EGGS_PARROT = forge("eggs/parrot");
 	public static final TagKey<Item> DOUGH_CORN = forge("dough/corn");
 	public static final TagKey<Item> DOUGH_NUT = forge("dough/nut");
-	public static final TagKey<Item> BREAD = forge("bread");
-	public static final TagKey<Item> BREAD_WHEAT = forge("bread/wheat");
 	public static final TagKey<Item> BREAD_CORN = forge("bread/corn");
 	public static final TagKey<Item> BREAD_SLICES = forge("bread_slices");
 	public static final TagKey<Item> BREAD_SLICES_TOASTED = forge("bread_slices/toasted");
@@ -134,7 +135,7 @@ public class DelightfulItemTags {
 	public static final TagKey<Item> GUMMIES = Util.it("collectorsreap", "gummies");
 
 	// Nethers Delight
-	public static final TagKey<Item> MEAL_ITEM = Util.it("nethersdelight", "meal_item");
+	public static final TagKey<Item> MEAL_ITEM = Util.it(Mods.ND, "meal_item");
 
 	// Ecologics
 	public static final TagKey<Item> COOKED_PRICKLY_PEAR = Util.it("ecologics", "cooked_prickly_pear");